improve docs & packaging
