+++ changelog.md
@@ -1,0 +1,9 @@
+# Changelog for `build-env`
+
+## Version 1.0.0.0 (2022-12-XX)
+
+- First Hackage release.
+
+## Version 0.1.0.0 (2022-11-16)
+
+- Initial proof of concept.
